Move ModelViews to ModelView folder
