+++ Dinamic/Model/ModelView/PartidaImpuestoResumen.php
@@ -1,0 +1,10 @@
+<?php namespace FacturaScripts\Dinamic\Model\ModelView;
+
+/**
+ * Class created by Core/Base/PluginManager
+ * @package FacturaScripts\Dinamic\Model\ModelView
+ * @author Carlos García Gómez <carlos@facturascripts.com>
+ */
+class PartidaImpuestoResumen extends \FacturaScripts\Core\Model\ModelView\PartidaImpuestoResumen
+{
+}
