--- Dinamic/Model/PartidaImpuesto.php
@@ -1,10 +1,0 @@
-<?php namespace FacturaScripts\Dinamic\Model;
-
-/**
- * Class created by Core/Base/PluginManager
- * @package FacturaScripts\Dinamic\Model
- * @author Carlos García Gómez <carlos@facturascripts.com>
- */
-class PartidaImpuesto extends \FacturaScripts\Core\Model\PartidaImpuesto
-{
-}
